Added viewer redirects and file breadcrumb
